Separated line component separation out of GlsParser
It's now in a new class, `LineComponentSeparator`.
